Remove community calendar link
Following the public announcement of MicroShift, it will be critical to establish a community framework suited to public interest.  The community meeting solved this ad hoc but is largely unmanaged and lacking clear purpose.  I suggest we redirect all communications to the public slack channel until the MicroShift is able to dedicate resources to community engagement, designing a governance framework, etc.